Add Rector
This will help code get automagically get upgraded saving a lot of time while doing package upgrades to newer PHP versions. To assist with that a RectoConfig class has been added to this package that will cover 80%+ of all my packages needs.

The running of Rector is already visible in pr this